kafka-producer/requirements.txt
pyspark[sql,pandas_on_spark,connect]==3.5.4
kafka-python==2.0.2
pandas==1.5.3
numpy==1.23.5
fastavro==1.7.3
pytest==7.2.0
python-dotenv==0.21.1
notebook==6.5.4
ipykernel==6.21.0
matplotlib==3.6.3
seaborn==0.12.2
plotly==5.13.1
flask==2.2.3
flask-cors==3.0.10
